fix conflict, complete part 5

+++ src/main/java/org/example/sem_backend/modules/room_module/domain/entity/Room.java
@@ -20,7 +20,7 @@ import java.util.Set;
 public class Room {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int uniqueId;
+    private Long uniqueId;
 
     @NotBlank(message = "Tên phòng không được để trống")
     @Column(unique = true)
